Add inline svg element output
The purpose of the addition is to allow styling svg with .css classes. This is not possible if the svg is displayed as <img src="logo.svg">
The change adds new switch keyword (svg:) to icon tags, which allows output of inline svg elements as icon.
Change adds string sanitization of svg string to remove <script> tags. Most of string transformations are based on regex matching. I would appreciate help with code edit to add a toggle/switch to Quasar options object to opt in to letting inline svg output. Or possibly this could be an optional plugin.